move supabase to api folder, and fix provider import in toggle
